Update patch set 1
Patch Set 1: Code-Review-1

I find no real need to parametrize it (not against it) but I am for defaulting to use of 'ssh-add -L' result for loading user keys.

This is safe as it does not require exposing private keys.


PS. Downvote only because we need better default, which is likely to make the need for parameters much less likely to be needed.

Patch-set: 1
Reviewer: Gerrit User 371 <371@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1