First steps to enable zuul-runner
This patch create all the pieces to start use zuul-runner.
Since the patches required to execute it locally are merged upstream
yet, this playbooks generate the zuul container with podman with the
patch #657955 to start the zuul-runner.
It also add a new job to test this playbook.
Next teps are to create another job that will in fact create the vm
to test the job.

Change-Id: I2891121d65104bf6bcbac7420514bb0d7ed12e11