Fix tox linters
Ansible-lint have an issue where it fails when you try to convert an
undefined value using to_nice_yaml or to_nice_json. This is undefined
variable is just for the linter, since in execution time, all these
variables are set properly. To avoid the failures in the linter, we set
the default to empty.
https://github.com/ansible/ansible-lint/issues/506 is the issue.

Change-Id: I9abe74ac6e85aa5367e637f67826964515ded31d