Update patch set 1
Patch Set 1:

This addition to the readme does contain value, but this is goings toward the incorrect path IMO. Zuul reproducer is the best place to manually launch locally our integration tests. There are two terms here that need to be stressed: manual, and integration. These are the only kind of test that we currently have: manual, integration and manual launch of integration. Having only these kinds of tests is not scalable, sustainable nor efficient. Integration test CANNOT be the only tests we use: they take too many resources and time, their goal is mainly test of tripleo, and they are completely insufficient to test our code. Introducing functional tests in our code is the correct path, for both short term and long term. Our roles handle different input sets, and produce different outcome, usually one per integration job. If what we run is only integration tests, we are bound to test only a small set of role paths each time, and you cannot run 100 integration jobs for all the different outcomes for the roles.
Functional tests usually take minutes, so we can exhaust all the possible paths in the same role in a very short time, without the need of a the whole support of the reproducer, because functional tests are small and simple enough to be run everywhere.

Patch-set: 1
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0