Make zuul writing comments about errors in configs
In order to provide better feedback about possible configuration errors
in zuul.yaml, it needs to be presented in the repo in advance.
And it needs to provide an empty list of check jobs.

Then, after the reproducer script submits its rendered zuul.yaml with
zuul_job under reproduction, zull will happily post about any
discovered errors in it right in the patch comments.

Change-Id: I2fedd5e7ddea2f88fd466f400d2695a25c28c556
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>