Do not omit package_mirror, if undefined
In the current form of ci_jobs_config.yaml, it cannot
template undefined package_mirror properly, since that's a vars
file.

So when package_mirror was not specified in launcher-playbook.yaml, do
not use |omit filter in ci_jobs_config.yaml, as that passes down the
wrong __omit_place_holder__ value.  Instead just never include it via
the vars file, and hope for the better (it should be passed in
elsewhere).

Change-Id: If940e85ce27711d7482b6cc8cf6fd3df30c772fd
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>