Reduce ansible usage in containers
Installing ansible in some of the containers to do some basic setup is
very time and cpu consuming. This change switches the playbooks for ssh
keys and waiting for gerrit & mysql to use a bash script rather than
ansible.  The gerrit config is still performed via ansible so the
timeouts needed to be raised in comparison to the ansible playbooks as
the shell scripts are invoked much sooner than the playbooks were due to
ansible pip install time.

Additionally adds a variable (`podman_pull`) to skip the podman pull
invocation before the pod create.

Change-Id: I926fa8629b9d8d9286f0d78e461dd3376745a92b